golangci-lint on each of the modules
This is necessary because of the way go handles listing packages in
repositories where more than one module is present.

More info can be found here: https://github.com/golang/go/wiki/Modules#what-are-multi-module-repositories

Also renamed the sub modules so they are inline with the name of the
base module.